refactor(경로): 수학적 경유지 생성기 삭제 및 경유지 생성기 패키지 재정의
